+++ app.go
@@ -21,7 +21,7 @@ import (
 	"time"
 )
 
-var version = "0.11.0"
+var version = "0.11.1"
 
 const (
 	maxCachedQueries = 512